Updated to return floor as int

+++ 0x00-python_variable_annotations/2-floor.py
@@ -2,11 +2,9 @@
 """Type-annotation function floor
 Takes in float argument and returns the
 floor of the float"""
-
-
 import math
 
 
-def floor(n: float) -> float:
+def floor(n: float) -> int:
     """Floor function"""
     return math.floor(n)
